RF: move both scale space classes to their own module
