+++ dipy/align/imaffine.py
@@ -9,103 +9,16 @@ from dipy.core.optimize import Optimizer
 import matplotlib.pyplot as plt
 from dipy.align.transforms import (Transform)
 from dipy.align.imwarp import (get_direction_and_spacings, ScaleSpace)
-
-class IsotropicScaleSpace(ScaleSpace):
-    def __init__(self, image, factors, sigmas,
-                 codomain_affine=None,
-                 input_spacing=None,
-                 mask0=False):
-        r"""
-        Example:
-        image = np.zeros(shape=(256, 256, 128))
-        factors = [4, 2, 1]
-        sigmas = [3, 1, 0]
-        codomain_affine = None
-        input_spacing = np.array([1., 1., 1.5])
-        mask0 = False
-        ss = IsotropicScaleSpace(image, factors, sigmas, codomain_affine,
-                                 input_spacing, mask0)
-        shape = ss.get_domain_shape(2)
-        """
-        self.dim = len(image.shape)
-        self.num_levels = len(factors)
-        if len(sigmas) != self.num_levels:
-            raise ValueError("sigmas and factors must have the same length")
-        input_size = np.array(image.shape)
-        if mask0:
-            mask = np.asarray(image>0, dtype=np.int32)
-
-        #normalize input image to [0,1]
-        img = (image - image.min())/(image.max() - image.min())
-        if mask0:
-            img *= mask
-
-        #The properties are saved in separate lists. Insert input image
-        #properties at the first level of the scale space
-        self.images = [img.astype(floating)]
-        self.domain_shapes = [input_size.astype(np.int32)]
-        if input_spacing is None:
-            input_spacing = np.ones((self.dim,), dtype = np.int32)
-        self.spacings = [input_spacing]
-        self.scalings = [np.ones(self.dim)]
-        self.affines = [codomain_affine]
-        self.sigmas = [np.ones(self.dim) * sigmas[self.num_levels - 1]]
-
-        if codomain_affine is not None:
-            self.affine_invs = [npl.inv(codomain_affine)]
-        else:
-            self.affine_invs = [None]
-
-        #compute the rest of the levels
-        min_index = np.argmin(input_spacing)
-        min_spacing = input_spacing[min_index]
-        for i in range(1, self.num_levels):
-            factor = factors[self.num_levels - 1 - i]
-            shrink_factors = np.zeros(self.dim)
-            new_spacing = np.zeros(self.dim)
-            shrink_factors[min_index] = factor
-            new_spacing[min_index] = input_spacing[min_index] * factor
-            for j in range(self.dim):
-                if j != min_index:
-                    #Select the factor that maximizes isotropy
-                    shrink_factors[j] = factor
-                    new_spacing[j] = input_spacing[j] * factor
-                    min_diff = np.abs(new_spacing[j] - new_spacing[min_index])
-                    for f in range(1, factor):
-                        diff = np.abs(input_spacing[j] * f - new_spacing[min_index])
-                        if diff < min_diff:
-                            shrink_factors[j] = f
-                            new_spacing[j] = input_spacing[j] * f
-                            min_diff = diff
-
-            extended = np.append(shrink_factors, [1])
-            if not codomain_affine is None:
-                affine = codomain_affine.dot(np.diag(extended))
-            else:
-                affine = np.diag(extended)
-            output_size = (input_size / shrink_factors).astype(np.int32)
-            new_sigmas = np.ones(self.dim) * sigmas[self.num_levels - i - 1]
-
-            #filter along each direction with the appropriate sigma
-            filtered = ndimage.filters.gaussian_filter(image, new_sigmas)
-            filtered = ((filtered - filtered.min())/
-                       (filtered.max() - filtered.min()))
-            if mask0:
-                filtered *= mask
-
-            #Add current level to the scale space
-            self.images.append(filtered.astype(floating))
-            self.domain_shapes.append(output_size)
-            self.spacings.append(new_spacing)
-            self.scalings.append(shrink_factors)
-            self.affines.append(affine)
-            self.affine_invs.append(npl.inv(affine))
-            self.sigmas.append(new_sigmas)
+from dipy.align.scalespace import IsotropicScaleSpace
 
 
 class MattesMIMetric(MattesBase):
     def __init__(self, nbins=32, sampling_proportion=30):
         r""" Initializes an instance of the Mattes MI metric
+
+        This class implements the methods required by Optimizer to drive the
+        registration process by making calls to the low level methods defined
+        in MattesBase.
 
         Parameters
         ----------
@@ -118,14 +31,14 @@ class MattesMIMetric(MattesBase):
         super(MattesMIMetric, self).__init__(nbins)
         self.sampling_proportion = sampling_proportion
 
-    def setup(self, transform, static, moving, moving_spacing, static_aff=None,
-              moving_aff=None, smask=None, mmask=None, prealign=None,
-              precondition=False):
+    def setup(self, transform, static, moving, moving_spacing,
+              static_grid2space=None, moving_grid2space=None, prealign=None):
         r""" Prepares the metric to compute intensity densities and gradients
 
         The histograms will be setup to compute probability densities of
-        intensities within the minimum and maximum values of static and moving
-        within the given masks
+        intensities within the minimum and maximum values of `static` and
+        `moving`
+
 
         Parameters
         ----------
@@ -138,18 +51,10 @@ class MattesMIMetric(MattesBase):
             moving image
         moving_spacing : array, shape (3,) or (2,)
             the spacing between neighboring voxels along each dimension
-        static_affine : array (dim+1, dim+1)
+        static_grid2space : array (dim+1, dim+1)
             the grid-to-space transform of the static image
-        moving_affine : array (dim+1, dim+1)
+        moving_grid2space : array (dim+1, dim+1)
             the grid-to-space transform of the moving image
-        smask : array, shape (S, R, C) or (R, C)
-            the mask indicating the voxels of interest within the static image.
-            If None, ones_like(smask) will be used as mask (all voxels will be
-            taken).
-        mmask : array, shape (S', R', C') or (R', C')
-            the mask indicating the voxels of interest within the moving image
-            If None, ones_like(smask) will be used as mask (all voxels will be
-            taken).
         prealign : array, shape (dim+1, dim+1)
             the pre-aligning matrix (an affine transform) that roughly aligns
             the moving image towards the static image. If None, no pre-alignment
@@ -158,24 +63,18 @@ class MattesMIMetric(MattesBase):
             instead of manually warping the moving image and provide None or
             identity as prealign. This way, the metric avoids performing more
             than one interpolation.
-        precondition : boolean
-            if True, the gradient's components are scaled attempting to make the
-            levelsets of the objective function as isotropic as possible
         """
         self.dim = len(static.shape)
         self.transform = transform
         self.static = np.array(static).astype(np.float64)
         self.moving = np.array(moving).astype(np.float64)
-        self.static_aff = static_aff
-        self.static_aff_inv = npl.inv(static_aff)
-        self.moving_aff = moving_aff
-        self.moving_aff_inv = npl.inv(moving_aff)
+        self.static_grid2space = static_grid2space
+        self.static_space2grid = npl.inv(static_grid2space)
+        self.moving_grid2space = moving_grid2space
+        self.moving_space2grid = npl.inv(moving_grid2space)
         self.moving_spacing = moving_spacing
-        self.smask = smask
-        self.mmask = mmask
         self.prealign = prealign
         self.param_scales = None
-        self.precondition = precondition
 
         T = np.eye(self.dim + 1)
         if self.prealign is not None:
@@ -187,25 +86,25 @@ class MattesMIMetric(MattesBase):
             self.interp_method = vf.interpolate_scalar_3d
 
         if self.sampling_proportion is None:
-            self.warped = aff_warp(self.static, self.static_aff, self.moving,
-                                   self.moving_aff, T).astype(np.float64)
+            self.warped = aff_warp(self.static, self.static_grid2space, self.moving,
+                                   self.moving_grid2space, T).astype(np.float64)
             self.samples = None
             self.ns = 0
         else:
             self.warped = None
             k = 100/self.sampling_proportion
             shape = np.array(static.shape, dtype=np.int32)
-            samples = np.array(sample_domain_regular(k, shape, static_aff))
+            samples = np.array(sample_domain_regular(k, shape, static_grid2space))
             ns = samples.shape[0]
             # Add a column of ones (homogeneous coordinates)
             samples = np.hstack((samples, np.ones(ns)[:,None]))
             # Sample the static image
-            points_on_static = (self.static_aff_inv.dot(samples.T).T)[...,:self.dim]
+            points_on_static = (self.static_space2grid.dot(samples.T).T)[...,:self.dim]
             static_vals, inside = self.interp_method(self.static.astype(np.float32),
                                                      points_on_static)
             static_vals = np.array(static_vals, dtype=np.float64)
             # Sample the moving image
-            sp_to_moving = self.moving_aff_inv.dot(T)
+            sp_to_moving = self.moving_space2grid.dot(T)
             points_on_moving = (sp_to_moving.dot(samples.T).T)[...,:self.dim]
             moving_vals, inside = self.interp_method(self.moving.astype(np.float32),
                                                      points_on_moving)
@@ -215,21 +114,17 @@ class MattesMIMetric(MattesBase):
             self.static_vals = static_vals
             self.moving_vals = moving_vals
 
-        if self.mmask is not None:
-            self.wmask = aff_warp(self.static, self.static_aff, self.mmask,
-                                  self.moving_aff, T, True).astype(np.int32)
-        else:
-            self.wmask = None
-
-        MattesBase.setup(self, self.static, self.moving, self.smask, self.wmask)
+        MattesBase.setup(self, self.static, self.moving)
 
     def _update_dense(self, xopt, update_gradient = True):
-        r""" Updates the marginal and joint distributions and the joint gradient
+        r""" Updates marginal and joint distributions and the joint gradient
 
-        The distributions and the gradient of the joint distribution are
-        updated according to the static and warped images. The warped image
-        is precisely the moving image after transforming it by the transform
-        defined by the xopt parameters.
+        The distributions are updated according to the static and warped
+        images. The warped image is precisely the moving image after
+        transforming it by the transform defined by the xopt parameters.
+
+        The gradient of the joint PDF is computed only if update_gradient
+        is True.
 
         Parameters
         ----------
@@ -237,6 +132,9 @@ class MattesMIMetric(MattesBase):
             the parameter vector of the transform currently used by the metric
             (the transform name is provided when self.setup is called), n is
             the number of parameters of the transform
+        update_gradient : Boolean
+            if true, the gradient of the joint PDF will also be computed,
+            otherwise, only the marginal and joint PDFs will be computed
         """
         # Get the matrix associated to the xopt parameter vector
         T = self.transform.param_to_matrix(xopt)
@@ -244,35 +142,28 @@ class MattesMIMetric(MattesBase):
             T = T.dot(self.prealign)
 
         # Warp the moving image
-        self.warped = aff_warp(self.static, self.static_aff, self.moving,
-                               self.moving_aff, T).astype(np.float64)
-
-        if self.mmask is not None:
-            # Get the warped mask.
-            self.wmask = aff_warp(self.static, self.static_aff, self.mmask,
-                                  self.moving_aff, T, True).astype(np.int32)
-        else:
-            self.wmask = None
+        self.warped = aff_warp(self.static, self.static_grid2space, self.moving,
+                               self.moving_grid2space, T).astype(np.float64)
 
         # Update the joint and marginal intensity distributions
-        self.update_pdfs_dense(self.static, self.warped, self.smask, self.wmask)
+        self.update_pdfs_dense(self.static, self.warped, None, None)
 
         # Compute the gradient of the joint PDF w.r.t. parameters
         if update_gradient:
-            if self.static_aff is None:
+            if self.static_grid2space is None:
                 grid_to_space = T
             else:
-                grid_to_space = T.dot(self.static_aff)
+                grid_to_space = T.dot(self.static_grid2space)
 
             # Compute the gradient of the moving image at the current transform
-            grid_to_space = T.dot(self.static_aff)
-            self.grad_w, inside = vf.gradient(self.moving.astype(floating), self.moving_aff_inv,
+            grid_to_space = T.dot(self.static_grid2space)
+            self.grad_w, inside = vf.gradient(self.moving.astype(floating), self.moving_space2grid,
                                       self.moving_spacing, self.static.shape, grid_to_space)
 
             # Update the gradient of the metric
             self.update_gradient_dense(xopt, self.transform, self.static,
                                        self.warped, grid_to_space, self.grad_w,
-                                       self.smask, self.wmask)
+                                       None, None)
 
         # Evaluate the mutual information and its gradient
         # The results are in self.metric_val and self.metric_grad
@@ -283,11 +174,13 @@ class MattesMIMetric(MattesBase):
     def _update_sparse(self, xopt, update_gradient = True):
         r""" Updates the marginal and joint distributions and the joint gradient
 
-        The distributions and the gradient of the joint distribution are
-        updated according to the samples taken from the static and moving images.
-        The samples are points in physical space, so the static intensities are
-        are always the same, but the corresponding points in the moving image
-        depend on the transform defined by xopt.
+        The distributions are updated according to the samples taken from the
+        static and moving images. The samples are points in physical space,
+        so the static intensities are always the same, but the corresponding
+        points in the moving image depend on the transform defined by xopt.
+
+        The gradient of the joint PDF is computed only if update_gradient
+        is True.
 
         Parameters
         ----------
@@ -295,6 +188,9 @@ class MattesMIMetric(MattesBase):
             the parameter vector of the transform currently used by the metric
             (the transform name is provided when self.setup is called), n is
             the number of parameters of the transform
+        update_gradient : Boolean
+            if true, the gradient of the joint PDF will also be computed,
+            otherwise, only the marginal and joint PDFs will be computed
         """
         # Get the matrix associated to the xopt parameter vector
         T = self.transform.param_to_matrix(xopt)
@@ -302,7 +198,7 @@ class MattesMIMetric(MattesBase):
             T = T.dot(self.prealign)
 
         # Sample the moving image
-        sp_to_moving = self.moving_aff_inv.dot(T)
+        sp_to_moving = self.moving_space2grid.dot(T)
         points_on_moving = (sp_to_moving.dot(self.samples.T).T)[...,:self.dim]
         self.moving_vals, inside = self.interp_method(self.moving.astype(np.float32),
                                                       points_on_moving)
@@ -330,7 +226,8 @@ class MattesMIMetric(MattesBase):
 
 
     def distance(self, xopt):
-        r""" Numeric value of the metric evaluated at the given parameters
+        r""" Numeric value of the MI metric evaluated at the given parameters
+
         Parameters
         ----------
         xopt : array, shape (n,)
@@ -346,6 +243,7 @@ class MattesMIMetric(MattesBase):
 
     def gradient(self, xopt):
         r""" Numeric value of the metric's gradient at the given parameters
+
         Parameters
         ----------
         xopt : array, shape (n,)
@@ -357,14 +255,12 @@ class MattesMIMetric(MattesBase):
             self._update_dense(xopt, True)
         else:
             self._update_dense(xopt, True)
-        if self.precondition:
-            return self.metric_grad / self.param_scales
-        else:
-            return self.metric_grad.copy()
+        return self.metric_grad.copy()
+
 
 
     def value_and_gradient(self, xopt):
-        r""" Numeric value of the metric and its gradient at the given parameter
+        r""" Numeric value of the metric and its gradient at given parameters
 
         Parameters
         ----------
@@ -377,18 +273,13 @@ class MattesMIMetric(MattesBase):
             self._update_dense(xopt, True)
         else:
             self._update_sparse(xopt, True)
-        print(">>",self.metric_val)
-        #print(">>",self.metric_val, self.param_scales, self.metric_grad.copy() / self.param_scales)
-        if self.precondition:
-            return self.metric_val, self.metric_grad / self.param_scales
-        else:
-            return self.metric_val, self.metric_grad.copy()
+        return self.metric_val, self.metric_grad.copy()
 
 
 class AffineRegistration(object):
     def __init__(self,
                  metric=None,
-                 method = 'CGGS',
+                 method = 'BFGS',
                  level_iters=None,
                  opt_tol=1e-5,
                  ss_sigma_factor=1.0,
@@ -401,6 +292,8 @@ class AffineRegistration(object):
         ----------
         metric : object
             an instance of a metric
+        method : string
+            optimization method to be used
         level_iters : list
             the number of iterations at each level of the Gaussian pyramid.
             level_iters[0] corresponds to the finest level, level_iters[n-1] the
@@ -411,6 +304,12 @@ class AffineRegistration(object):
             parameter of the scale-space smoothing kernel. For example, the
             std. dev. of the kernel will be factor*(2^i) in the isotropic case
             where i = 0, 1, ..., n_scales is the scale
+        factors : list of floats
+            custom scale factors to build the scale space (one factor for each
+            scale)
+        sigmas : list of floats
+            custom smoothing parameter to build the scale space (one parameter
+            for each scale)
         options : None or dict,
             extra optimization options.
         """
@@ -441,7 +340,7 @@ class AffineRegistration(object):
 
 
     def _init_optimizer(self, static, moving, transform, x0,
-                        static_affine, moving_affine, prealign):
+                        static_grid2space, moving_grid2space, prealign):
         r"""Initializes the registration optimizer
 
         Initializes the optimizer by computing the scale space of the input
@@ -464,14 +363,19 @@ class AffineRegistration(object):
             parameters from which to start the optimization. If None, the
             optimization will start at the identity transform. n is the
             number of parameters of the specified transformation.
-        static_affine: array, shape (dim+1, dim+1)
+        static_grid2space: array, shape (dim+1, dim+1)
             the voxel-to-space transformation associated to the static image
-        moving_affine: array, shape (dim+1, dim+1)
+        moving_grid2space: array, shape (dim+1, dim+1)
             the voxel-to-space transformation associated to the moving image
-        prealign: array, shape (dim+1, dim+1)
-            the affine transformation (operating on the physical space)
-            pre-aligning the moving image towards the static
-
+        prealign: string, or matrix, or None
+            If string:
+                'mass': align centers of gravity
+                'origins': align physical coordinates of voxel (0,0,0)
+                'centers': align physical coordinates of central voxels
+            If matrix:
+                array, shape (dim+1, dim+1)
+            If None:
+                Start from identity
         """
         self.dim = len(static.shape)
         self.transform = transform
@@ -483,21 +387,21 @@ class AffineRegistration(object):
         if prealign is None:
             self.prealign = np.eye(self.dim + 1)
         elif prealign == 'mass':
-            self.prealign = aff_centers_of_mass(static, static_affine, moving,
-                                                moving_affine)
+            self.prealign = aff_centers_of_mass(static, static_grid2space, moving,
+                                                moving_grid2space)
         elif prealign == 'origins':
-            self.prealign = aff_origins(static, static_affine, moving,
-                                        moving_affine)
+            self.prealign = aff_origins(static, static_grid2space, moving,
+                                        moving_gris2space)
         elif prealign == 'centers':
-            self.prealign = aff_geometric_centers(static, static_affine, moving,
-                                                  moving_affine)
+            self.prealign = aff_geometric_centers(static, static_grid2space, moving,
+                                                  moving_grid2space)
         else:
             self.prealign = prealign
         #Extract information from the affine matrices to create the scale space
         static_direction, static_spacing = \
-            get_direction_and_spacings(static_affine, self.dim)
+            get_direction_and_spacings(static_grid2space, self.dim)
         moving_direction, moving_spacing = \
-            get_direction_and_spacings(moving_affine, self.dim)
+            get_direction_and_spacings(moving_grid2space, self.dim)
 
         static = (static - static.min())/(static.max() - static.min())
         moving = (moving - moving.min())/(moving.max() - moving.min())
@@ -505,26 +409,42 @@ class AffineRegistration(object):
 
         if self.use_isotropic:
             self.moving_ss = IsotropicScaleSpace(moving, self.factors, self.sigmas,
-                                    moving_affine, moving_spacing, False)
+                                    moving_grid2space, moving_spacing, False)
 
             self.static_ss = IsotropicScaleSpace(static, self.factors, self.sigmas,
-                                        static_affine, static_spacing, False)
+                                        static_grid2space, static_spacing, False)
         else:
-            self.moving_ss = ScaleSpace(moving, self.levels, moving_affine,
+            self.moving_ss = ScaleSpace(moving, self.levels, moving_grid2space,
                                     moving_spacing, self.ss_sigma_factor,
                                     False)
 
-            self.static_ss = ScaleSpace(static, self.levels, static_affine,
+            self.static_ss = ScaleSpace(static, self.levels, static_grid2space,
                                     static_spacing, self.ss_sigma_factor,
                                     False)
 
 
-    def optimize(self, static, moving, transform, x0, static_affine=None,
-                 moving_affine=None, smask=None, mmask=None, prealign=None):
+    def optimize(self, static, moving, transform, x0, static_grid2space=None,
+                 moving_grid2space=None, prealign=None):
         r'''
         Parameters
         ----------
+        static: array, shape (S, R, C) or (R, C)
+            the image to be used as reference during optimization.
+        moving: array, shape (S, R, C) or (R, C)
+            the image to be used as "moving" during optimization. It is
+            necessary to pre-align the moving image to ensure its domain
+            lies inside the domain of the deformation fields. This is assumed to
+            be accomplished by "pre-aligning" the moving image towards the
+            static using an affine transformation given by the 'prealign' matrix
         transform: string
+        x0: array, shape (n,)
+            parameters from which to start the optimization. If None, the
+            optimization will start at the identity transform. n is the
+            number of parameters of the specified transformation.
+        static_grid2space: array, shape (dim+1, dim+1)
+            the voxel-to-space transformation associated to the static image
+        moving_grid2space: array, shape (dim+1, dim+1)
+            the voxel-to-space transformation associated to the moving image
         prealign: string, or matrix, or None
             If string:
                 'mass': align centers of gravity
@@ -535,17 +455,14 @@ class AffineRegistration(object):
             If None:
                 Start from identity
         '''
-        self._init_optimizer(static, moving, transform, x0, static_affine,
-                             moving_affine, prealign)
+        self._init_optimizer(static, moving, transform, x0, static_grid2space,
+                             moving_grid2space, prealign)
         del prealign # now we must refer to self.prealign
 
         # Multi-resolution iterations
-        original_static_affine = self.static_ss.get_affine(0)
-        original_moving_affine = self.moving_ss.get_affine(0)
+        original_static_grid2space = self.static_ss.get_affine(0)
+        original_moving_grid2space = self.moving_ss.get_affine(0)
         original_moving_spacing = self.moving_ss.get_spacing(0)
-
-        original_smask = smask
-        original_mmask = mmask
 
         for level in range(self.levels - 1, -1, -1):
             self.current_level = level
@@ -555,60 +472,31 @@ class AffineRegistration(object):
             # Resample the smooth static image to the shape of this level
             smooth_static = self.static_ss.get_image(level)
             current_static_shape = self.static_ss.get_domain_shape(level)
-            current_static_aff = self.static_ss.get_affine(level)
+            current_static_grid2space = self.static_ss.get_affine(level)
 
             current_static = aff_warp(tuple(current_static_shape),
-                                      current_static_aff, smooth_static,
-                                      original_static_affine, None, False)
-            if original_smask is not None:
-                current_smask = aff_warp(tuple(current_static_shape),
-                                         current_static_aff, original_smask,
-                                         original_static_affine, None, True)
-            else:
-                current_smask = None
+                                      current_static_grid2space, smooth_static,
+                                      original_static_grid2space, None, False)
 
             # The moving image is full resolution
-            current_moving_aff = original_moving_affine
+            current_moving_grid2space = original_moving_grid2space
             current_moving_spacing = self.moving_ss.get_spacing(level)
 
             current_moving = self.moving_ss.get_image(level)
-            current_mmask = original_mmask
 
             # Prepare the metric for iterations at this resolution
-            if self.method == 'CGGS':
-                precondition = True
-            else:
-                precondition = False
             self.metric.setup(transform, current_static, current_moving,
-                              current_moving_spacing, current_static_aff, current_moving_aff,
-                              current_smask, current_mmask, self.prealign,
-                              precondition)
-
-            if precondition:
-                grid_to_space = self.prealign.dot(current_static_aff)
-                scales = estimate_param_scales(self.transform_type, self.dim,
-                                               current_static.shape,
-                                               grid_to_space)
-                self.metric.param_scales = scales
+                              current_moving_spacing, current_static_grid2space,
+                              current_moving_grid2space, self.prealign)
 
             #optimize this level
             if self.options is None:
                 self.options = {'gtol':1e-4, 'maxiter': max_iter, 'disp':False}
-            if self.method == 'CGGS':
-                domain_shape = current_static.shape
-                domain_affine = self.prealign.dot(current_static_aff)
 
-                xopt = nonlinear_cg(self.transform_type, self.dim, domain_shape,
-                                    domain_affine,
-                                    self.metric.distance,
-                                    self.metric.gradient,
-                                    self.metric.value_and_gradient,
-                                    self.x0, self.options)
-            else:
-                opt = Optimizer(self.metric.value_and_gradient, self.x0,
-                                method=self.method, jac = True,
-                                options=self.options)
-                xopt = opt.xopt
+            opt = Optimizer(self.metric.value_and_gradient, self.x0,
+                            method=self.method, jac = True,
+                            options=self.options)
+            xopt = opt.xopt
 
             # Update prealign matrix with optimal parameters
             T = self.transform.param_to_matrix(xopt)
@@ -622,70 +510,7 @@ class AffineRegistration(object):
         return self.prealign
 
 
-def estimate_param_scales(transform_type, dim, domain_shape, domain_affine):
-    r""" Estimate the parameter scales of the given affine transform
-
-    If we vary only one of the parameters of the parameter vector p of the
-    given parametric transformation T (where p corresponds to the parameters of
-    the identity transform), this defines a trajectory T(p[i]; x0), which
-    transforms an initial point x0 according to the parameter p[i]. The
-    magnitude of the tangent vector of that trajectory is a measure of its
-    'velocity', which tells us how 'sensitive' the transformation is to each of
-    its parameters at the specific point x0. Thus, we define the 'scale' of a
-    parameter as the maximum, over all possible starting points x0, of the
-    aforementioned tangent vector's squared norm.
-
-    Parameters
-    ----------
-    transform_type : int
-        the type of the transformation (use transform_type dictionary from
-        the transforms module to map transformation name to int)
-    dim : int (either 2 or 3)
-        the domain dimension of the transformation (either 2 or 3)
-    domain_shape : array, shape (dim,)
-        the shape of the grid on which the transform is applied
-    domain_affine : array, shape (dim + 1, dim + 1)
-        the grid-to-space transform associated to the grid of shape domain_shape
-    """
-    h = 0.01
-    n = number_of_parameters(transform_type, dim)
-    theta = np.empty(n)
-    X = np.empty((2 ** dim, dim + 1)) # All 2^dim corners of the grid
-    T = np.ndarray((dim + 1, dim + 1))
-
-    # Generate all corners of the given domain
-    X[:, dim] = 1 # Homogeneous coordinate
-    for i in range(2 ** dim):
-        ii = i
-        for j in range(dim):
-            if (ii % 2) == 0:
-                X[i, j] = 0
-            else:
-                X[i, j] = domain_shape[j] - 1
-            ii = ii // 2
-
-    # Transform grid points to physical space
-    if domain_affine is not None:
-        X = X.dot(domain_affine.transpose())
-
-    # Compute the scale of each parameter
-    scales = np.zeros(n)
-    for i in range(n):
-        get_identity_parameters(transform_type, dim, theta)
-        theta[i] += h
-        param_to_matrix(transform_type, dim, theta, T)
-        transformed = X.dot(T.transpose())
-        sq_norms = np.sum((transformed - X) ** 2, 1)
-        max_shift_sq = sq_norms.max()
-        scales[i] = max_shift_sq
-
-    # Avoid zero scales to prevent division by zero
-    scales[scales == 0] = scales[scales > 0].min()
-    scales /= h * h
-    return scales
-
-
-def aff_warp(static, static_affine, moving, moving_affine, transform, nn=False):
+def aff_warp(static, static_grid2space, moving, moving_grid2space, transform, nn=False):
     r""" Warps the moving image towards the static using the given transform
 
     Parameters
@@ -693,11 +518,11 @@ def aff_warp(static, static_affine, moving, moving_affine, transform, nn=False):
     static: array, shape(S, R, C)
         static image: it will provide the grid and grid-to-space transform for
         the warped image
-    static_affine:
+    static_grid2space:
         grid-to-space transform associated to the static image
     moving: array, shape(S', R', C')
         moving image
-    moving_affine:
+    moving_grid2space:
         grid-to-space transform associated to the moving image
 
     Returns
@@ -723,29 +548,29 @@ def aff_warp(static, static_affine, moving, moving_affine, transform, nn=False):
         elif dim == 3:
             warp_method = vf.warp_3d_affine
 
-    m_aff_inv = npl.inv(moving_affine)
+    m_space2grid = npl.inv(moving_grid2space)
     if transform is None:
-        composition = m_aff_inv.dot(static_affine)
+        composition = m_space2grid.dot(static_grid2space)
     else:
-        composition = m_aff_inv.dot(transform.dot(static_affine))
+        composition = m_space2grid.dot(transform.dot(static_grid2space))
 
     warped = warp_method(input, shape, composition)
 
     return np.array(warped)
 
 
-def aff_centers_of_mass(static, static_affine, moving, moving_affine):
+def aff_centers_of_mass(static, static_grid2space, moving, moving_grid2space):
     r""" Transformation to align the center of mass of the input images
 
     Parameters
     ----------
     static: array, shape(S, R, C)
         static image
-    static_affine: array, shape (4, 4)
+    static_grid2space: array, shape (4, 4)
         the voxel-to-space transformation of the static image
     moving: array, shape(S, R, C)
         moving image
-    moving_affine: array, shape (4, 4)
+    moving_grid2space: array, shape (4, 4)
         the voxel-to-space transformation of the moving image
 
     Returns
@@ -757,15 +582,15 @@ def aff_centers_of_mass(static, static_affine, moving, moving_affine):
     """
     dim = len(static.shape)
     c_static = ndimage.measurements.center_of_mass(np.array(static))
-    c_static = static_affine.dot(c_static+(1,))
+    c_static = static_grid2space.dot(c_static+(1,))
     c_moving = ndimage.measurements.center_of_mass(np.array(moving))
-    c_moving = moving_affine.dot(c_moving+(1,))
+    c_moving = moving_grid2space.dot(c_moving+(1,))
     transform = np.eye(dim + 1)
     transform[:dim,dim] = (c_moving - c_static)[:dim]
     return transform
 
 
-def aff_geometric_centers(static, static_affine, moving, moving_affine):
+def aff_geometric_centers(static, static_grid2space, moving, moving_grid2space):
     r""" Transformation to align the geometric center of the input images
 
     With "geometric center" of a volume we mean the physical coordinates of
@@ -775,11 +600,11 @@ def aff_geometric_centers(static, static_affine, moving, moving_affine):
     ----------
     static: array, shape(S, R, C)
         static image
-    static_affine: array, shape (4, 4)
+    static_grid2space: array, shape (4, 4)
         the voxel-to-space transformation of the static image
     moving: array, shape(S, R, C)
         moving image
-    moving_affine: array, shape (4, 4)
+    moving_grid2space: array, shape (4, 4)
         the voxel-to-space transformation of the moving image
 
     Returns
@@ -791,15 +616,15 @@ def aff_geometric_centers(static, static_affine, moving, moving_affine):
     """
     dim = len(static.shape)
     c_static = tuple((np.array(static.shape, dtype = np.float64))*0.5)
-    c_static = static_affine.dot(c_static+(1,))
+    c_static = static_grid2space.dot(c_static+(1,))
     c_moving = tuple((np.array(moving.shape, dtype = np.float64))*0.5)
-    c_moving = moving_affine.dot(c_moving+(1,))
+    c_moving = moving_grid2space.dot(c_moving+(1,))
     transform = np.eye(dim + 1)
     transform[:dim,dim] = (c_moving - c_static)[:dim]
     return transform
 
 
-def aff_origins(static, static_affine, moving, moving_affine):
+def aff_origins(static, static_grid2space, moving, moving_grid2space):
     r""" Transformation to align the origins of the input images
 
     With "origin" of a volume we mean the physical coordinates of
@@ -809,11 +634,11 @@ def aff_origins(static, static_affine, moving, moving_affine):
     ----------
     static: array, shape(S, R, C)
         static image
-    static_affine: array, shape (4, 4)
+    static_grid2space: array, shape (4, 4)
         the voxel-to-space transformation of the static image
     moving: array, shape(S, R, C)
         moving image
-    moving_affine: array, shape (4, 4)
+    moving_grid2space: array, shape (4, 4)
         the voxel-to-space transformation of the moving image
 
     Returns
@@ -824,8 +649,8 @@ def aff_origins(static, static_affine, moving, moving_affine):
         image
     """
     dim = len(static.shape)
-    c_static = static_affine[:dim, dim]
-    c_moving = moving_affine[:dim, dim]
+    c_static = static_grid2space[:dim, dim]
+    c_moving = moving_grid2space[:dim, dim]
     transform = np.eye(dim + 1)
     transform[:dim,dim] = (c_moving - c_static)[:dim]
     return transform
